requirements.txt
# PID Auto-Tuner Dependencies
openai>=1.0.0          # DeepSeek API (OpenAI-compatible interface)
pyserial>=3.5          # Serial communication
numpy>=1.24.0          # Numerical analysis
matplotlib>=3.7.0      # Visualization (Phase 3)
